implement download-ci for coverage outcomes too
