+++ src/bootstrap/src/ferrocene/mod.rs
@@ -15,8 +15,10 @@ pub(crate) mod test_variants;
 pub(crate) mod tool;
 
 use std::collections::HashMap;
-use std::path::Path;
+use std::path::{Path, PathBuf};
 use std::process::Command;
+
+use build_helper::git::get_closest_merge_commit;
 
 use crate::builder::Builder;
 use crate::core::config::{Config, TargetSelection};
@@ -157,4 +159,46 @@ fn uv_command(builder: &Builder<'_>) -> BootstrapCommand {
     }
 
     command
+}
+
+fn download_and_extract_ci_outcomes(builder: &Builder<'_>, kind: &str) -> PathBuf {
+    if builder.config.dry_run() {
+        return PathBuf::new();
+    }
+
+    let name = format!("{kind}-outcomes");
+    let prefix = "s3://ferrocene-ci-artifacts/ferrocene/dist";
+    let commit = get_closest_merge_commit(None, &builder.config.git_config(), &[])
+        .expect(&format!("failed to retrieve git commit for ferrocene.{name}=download-ci"));
+
+    let base = builder.out.join("cache").join("ferrocene").join(&name);
+    let extracted_dir = base.join("extracted");
+    let tarballs_dir = base.join("tarballs");
+
+    let commit_file = extracted_dir.join(".ferrocene-commit");
+    let tarball_file = tarballs_dir.join(&format!("{commit}.tar.xz"));
+
+    if !tarball_file.exists() {
+        builder.info(&format!("Downloading {kind} outcomes for commit {commit}"));
+        let version = builder.config.artifact_version_part(&commit);
+        let url = format!("{prefix}/{commit}/ferrocene-{name}-{version}.tar.xz");
+        builder.create_dir(&tarballs_dir);
+        builder.config.download_file(
+            &url,
+            &tarball_file,
+            &format!("Could not download the {kind} outcomes."),
+        );
+    }
+
+    if !commit_file.exists() || builder.read(&commit_file) != commit {
+        builder.info(&format!("Extracting {kind} outcomes for commit {commit}"));
+        if extracted_dir.exists() {
+            builder.remove_dir(&extracted_dir);
+        }
+        builder.create_dir(&extracted_dir);
+        builder.config.unpack(&tarball_file, &extracted_dir, "");
+        std::fs::write(&commit_file, commit.as_bytes()).unwrap();
+    }
+
+    extracted_dir.join("share").join("ferrocene").join(&name)
 }
